requirements.txt
llama-cpp-python>=0.3.0
numpy>=1.24.0
fastapi>=0.100.0
uvicorn>=0.23.0
